fix(BotList): 修复停止未运行 bot 引发的错误 🛠️
- 结束进程时，尝试获取 pid，如果为空则表示 bot 未在运行，避免错误处理 🐛。

+++ src/Ui/BotListPage/BotWidget/__init__.py
@@ -217,9 +217,9 @@ class BotWidget(QWidget):
         """
         ## 停止按钮槽函数
         """
-        parent = psutil.Process(self.process.processId())
-        [child.kill() for child in parent.children(recursive=True)]
-        parent.kill()
+        if not (parent := psutil.Process(self.process.processId())):
+            [child.kill() for child in parent.children(recursive=True)]
+            parent.kill()
 
         self.isRun = False
         self.isLogin = False
